+++ cmd/ticket.go
@@ -22,8 +22,11 @@ import (
 	"path/filepath"
 	"runtime/pprof"
 	"strings"
+	"sync"
 	"syscall"
 
+	"github.com/panjf2000/ants/v2"
+	"github.com/rsvihladremio/ssdownloader/downloader"
 	"github.com/rsvihladremio/ssdownloader/link"
 	"github.com/rsvihladremio/ssdownloader/sendsafely"
 	"github.com/rsvihladremio/ssdownloader/zendesk"
@@ -32,6 +35,7 @@ import (
 )
 
 var useZendeskPassword bool
+var onlySendSafelyLinks bool
 
 // ticketCmd represents the ticket command
 var ticketCmd = &cobra.Command{
@@ -48,6 +52,7 @@ var ticketCmd = &cobra.Command{
 		`,
 	Args: cobra.ExactArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
+
 		if CpuProfile != "" {
 			f, err := os.Create(CpuProfile)
 			if err != nil {
@@ -73,6 +78,7 @@ var ticketCmd = &cobra.Command{
 			}
 
 		}
+		d := downloader.NewGenericDownloader(DownloadBufferSize)
 		password := ""
 		if useZendeskPassword {
 			fmt.Println("enter password:")
@@ -86,25 +92,104 @@ var ticketCmd = &cobra.Command{
 		}
 		zendeskApi := zendesk.NewZenDeskAPI(C.ZendeskEmail, password, C.ZendeskDomain, Verbose)
 		ticketId := args[0]
-		results, err := zendeskApi.GetTicketComents(ticketId)
+
+		results, err := zendeskApi.GetTicketComentsJSON(ticketId)
 		if err != nil {
 			log.Fatal(err)
 		}
-		for _, url := range results {
-			linkParts, err := link.ParseLink(url)
-			if err != nil {
-				log.Fatalf("unexpected error '%v' reading url '%v'", err, url)
+		urls, err := zendesk.GetLinksFromComments(results)
+		if err != nil {
+			log.Fatalf("unable parse comments with error '%v'", err)
+		}
+		p, err := ants.NewPool(DownloadThreads)
+		if err != nil {
+			log.Fatalf("cannot initialize thread pool due to error %v", err)
+		}
+
+		defer p.Release()
+		var wg sync.WaitGroup
+		for _, url := range urls {
+			if strings.HasPrefix(url, "https://sendsafely") {
+				linkParts, err := link.ParseLink(url)
+				if err != nil {
+					log.Fatalf("unexpected error '%v' reading url '%v'", err, url)
+				}
+				packageId := linkParts.PackageCode
+				wg.Add(1)
+				err = p.Submit(func() {
+					err := sendsafely.DownloadFilesFromPackage(d, packageId, linkParts.KeyCode, C, filepath.Join("tickets", ticketId), Verbose)
+					if err != nil {
+						log.Printf("error downloading %v", err)
+					}
+					wg.Done()
+				})
+				if err != nil {
+					log.Fatalf("cannot itialize sendsafely download due to error %v", err)
+				}
 			}
-			packageId := linkParts.PackageCode
-			err = sendsafely.DownloadFilesFromPackage(packageId, linkParts.KeyCode, C, filepath.Join("tickets", ticketId), Verbose)
+		}
+		if !onlySendSafelyLinks {
+			attachments, err := zendesk.GetAttachementsFromComments(results)
 			if err != nil {
 				log.Fatal(err)
 			}
+			for _, a := range attachments {
+				err = p.Submit(func() {
+					if err := DownloadNonSendSafelyLink(d, a, ticketId); err != nil {
+						log.Printf("WARN: error '%v' processing attachement %v skipping", err, a.FileName)
+					}
+					wg.Done()
+				})
+				if err != nil {
+					log.Fatalf("cannot itialize sendsafely download due to error %v", err)
+				}
+			}
 		}
+		wg.Wait()
 	},
+}
+
+func DownloadNonSendSafelyLink(d *downloader.GenericDownloader, a zendesk.Attachment, ticketId string) error {
+	if a.Deleted {
+		return fmt.Errorf("attachment '%v' from comment %v created on %v is marked as deleted, skipping", a.FileName, a.ParentCommentID, a.ParentCommentDate)
+	}
+	commentDir := fmt.Sprintf("%v_%v", a.ParentCommentDate.Format("2006-01-02T150405Z0700"), a.ParentCommentID)
+	downloadDir := filepath.Join(C.DownloadDir, "tickets", ticketId, "attachments", commentDir)
+	newFileName := filepath.Join(downloadDir, a.FileName)
+	if sendsafely.FileSizeMatches(newFileName, a.Size, Verbose) {
+		log.Printf("file '%v' already downloaded skipping", newFileName)
+		return nil
+	}
+	log.Printf("downloading attachment '%v'", newFileName)
+	if _, err := os.Stat(downloadDir); err != nil {
+		if os.IsNotExist(err) {
+			err = os.MkdirAll(downloadDir, 0700)
+			if err != nil {
+				return fmt.Errorf("unable to make dir %v due to error '%v", downloadDir, err)
+			}
+		} else {
+			return fmt.Errorf("unable to read dir %v due to error '%v", downloadDir, err)
+		}
+	}
+
+	if err := d.DownloadFile(newFileName, a.ContentURL); err != nil {
+		return fmt.Errorf("cannot download %v due to error %v, skipping", newFileName, err)
+	}
+	if !sendsafely.FileSizeMatches(newFileName, a.Size, Verbose) {
+		defer func() {
+			if err := os.Remove(newFileName); err != nil {
+				log.Printf("WARN: unexpected error '%v' trying to remove file '%v', remove manually", err, newFileName)
+			}
+		}()
+		return fmt.Errorf("file %v failed verification and did not meet expected size %v", newFileName, a.Size)
+	}
+	log.Printf("attachement '%v' is complete", newFileName)
+
+	return nil
 }
 
 func init() {
 	rootCmd.AddCommand(ticketCmd)
 	ticketCmd.Flags().BoolVarP(&useZendeskPassword, "zendesk-password", "p", false, "Use a password instead of an api key to authenticate against zendesk")
+	ticketCmd.Flags().BoolVar(&onlySendSafelyLinks, "sendsafely-only", false, "when true only sendsafely links will be downloaded")
 }