+++ zendesk/api.go
@@ -68,7 +68,7 @@ import (
 // 	  }
 // 	]
 //   }
-func (z *ZenDeskAPI) GetTicketComents(ticketId string) ([]string, error) {
+func (z *ZenDeskAPI) GetTicketComentsJSON(ticketId string) (string, error) {
 	url := fmt.Sprintf("https://%v.zendesk.com/api/v2/tickets/%v/comments.json", z.subDomain, ticketId)
 	auth := fmt.Sprintf("%v/token:%v", z.username, z.password)
 	base64Auth := base64.StdEncoding.EncodeToString([]byte(auth))
@@ -77,7 +77,7 @@ func (z *ZenDeskAPI) GetTicketComents(ticketId string) ([]string, error) {
 		SetHeader("Authorization", fmt.Sprintf("Basic %v", base64Auth)).
 		Get(url)
 	if err != nil {
-		return []string{}, fmt.Errorf("unable to read ticket comments with error '%v'", err)
+		return "", fmt.Errorf("unable to read ticket comments with error '%v'", err)
 	}
 	rawBody := r.Body()
 	if z.verbose {
@@ -88,12 +88,7 @@ func (z *ZenDeskAPI) GetTicketComents(ticketId string) ([]string, error) {
 			log.Printf("DEBUG: Ticket %v Comments Contents '%v'", ticketId, prettyJsonBuffer.String())
 		}
 	}
-	urls, err := GetSendSafelyLinksFromComments(string(rawBody))
-	if err != nil {
-		return []string{}, fmt.Errorf("unable parse comments with error '%v'", err)
-	}
-
-	return urls, nil
+	return string(rawBody), nil
 }
 
 type ZenDeskAPI struct {